refactor(structure): extract chapter inference boundaries
