+++ package.json
@@ -49,6 +49,7 @@
     "src/setup/",
     "src/scripts/",
     "src/styleguide/",
+    "src/structure/",
     "src/sync/",
     "src/tools/",
     "src/workflows/",
@@ -70,7 +71,7 @@
     "normalize:scene-characters": "node --experimental-sqlite src/scripts/normalize-scene-characters.mjs",
     "setup:openclaw-env": "sh src/scripts/setup-openclaw-env.sh",
     "release": "release-it",
-    "lint": "eslint *.js src/index.js src/core src/review-bundles src/runtime src/setup src/scripts src/styleguide src/sync src/tools src/workflows src/world",
+    "lint": "eslint *.js src/index.js src/core src/review-bundles src/runtime src/setup src/scripts src/structure src/styleguide src/sync src/tools src/workflows src/world",
     "guard:legacy-root-imports": "node src/scripts/check-legacy-root-imports.mjs",
     "docs": "node src/scripts/generate-tool-docs.mjs",
     "lint:metadata": "node src/scripts/lint-metadata.mjs",
